add transaction implementation

+++ backend-desapp-api/src/main/kotlin/ar/edu/unq/desapp/grupog/backenddesappapi/model/CryptoActive.kt
@@ -1,7 +1,7 @@
 package ar.edu.unq.desapp.grupog.backenddesappapi.model
 
 class CryptoActive(
-    var name: CryptoActiveName, val value: Long
+    var name: CryptoActiveName, val value: Double
 ) {
 
 }
